+++ simulation/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -e
+source /opt/ros/jazzy/setup.bash
+exec "$@"